PROCESSINGS: Process Feed Loading
